moved angle submission to interactive
